dastagiri-work1/code-files/requirements.txt
streamlit==1.28.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
prophet==1.1.4
plotly==5.15.0